Regroup and rename functions, classes and files
- Regroup skeleton reconstruction related functions into `skeleton.h`;
- Rename orientation_probability to orientation;
- Put BCC and CycleTracker to separate units;
- Rename BCC to BiconnectedComponent.
